feat(zsh): Add git stash with message alias and fzf file picker function
Adds a new alias `gstam` for quickly performing `git stash push -m` with a message. This streamlines the process of stashing changes with a descriptive message.

Also introduces a new `f` function that utilizes `fzf-tmux` to provide an interactive file picker. This allows users to easily select and retrieve file paths, making it convenient to pass them to other commands.